[Refactor] 유효한 로그인 값 상수로 지정한 후 import
