+++ js/utils/commons/value.trim.js
@@ -1,0 +1,5 @@
+const targetValue = function (event) {
+  return event.target.value.trim();
+};
+
+export { targetValue };
