+++ js/utils/constants.js
@@ -1,5 +1,8 @@
 // 유효한 email, password 값
-const validEmail = "test@codeit.com";
-const validPassword = "codeit101";
+export const validEmail = "test@codeit.com";
+export const validPassword = "codeit101";
 
-export { validEmail, validPassword };
+// 이메일 입력 값 확인
+export const SIGNIN_SUCCESS_VALUE = 1;
+export const PASSWORD_INVALID_VALUE = -1;
+export const NOT = 0;
